requirements.txt
selenium==3.141.0
webdriver_manager
goto-statement==1.2
requests
py-os
keras
dlib # requires cmake
libopencv
tensorflow
py-trello
trello
firebase
firebase_admin